Display the playing progress bar at the top of the bottom fragment
- Move the PlayProgress into it's own file
- Have the PlayProgress component listen to the playing state
- Fix the time status in the song info to be:
    currentTime / totalTime (remainingTime)
