+++ cmd/tchaik/ui/static/sass/variables.scss
@@ -2,6 +2,6 @@ $light-blue: #00D9FF;
 $left-column-width: 40px;
 $right-column-width: 350px;
 $filter-column-width: 250px;
-$bottom-row-height: 75px;
+$bottom-row-height: 70px;
 
 $light-background: #222;
